Move annotations to the constants module
